requirements.txt
django==5.2.4
djangorestframework==3.14.0
python-decouple==3.8
pillow==10.1.0
